Removed support for Ethereum presale wallets
