+++ cmd/gzond/main.go
@@ -215,7 +215,6 @@ func init() {
 		dumpGenesisCommand,
 		// See accountcmd.go:
 		accountCommand,
-		walletCommand,
 		// See consolecmd.go:
 		consoleCommand,
 		attachCommand,
